Ejecutable-luma-main/requirements.txt
requests
bs4
numpy
pandas
matplotlib
scikit-image
scipy
numba
argparse
opencv-python
PyWavelets
lxml
sqlalchemy
psycopg2
python-dotenv